🎨 sr pusher as kernel
